+++ src/pages/admin/AdminLayout.tsx
@@ -7,8 +7,8 @@ import {
   XMarkIcon,
 } from '@heroicons/react/24/outline'
 import { Fragment, FunctionComponent, PropsWithChildren, useState } from 'react'
-import Footer from '../../components/BaseFooter'
-import Logo from '../../components/Logo'
+import Footer from '@/components/BrandFooter'
+import Logo from '@/components/Logo'
 
 const navigation = [
   {
